requirements.txt
flask==3.1.3
anthropic==0.86.0
python-dotenv==1.2.2
gunicorn==23.0.0
